fixed a bug

+++ montezuma/main.py
@@ -281,9 +281,9 @@ for t in range(max_frames):
 			x = torch.Tensor(x).type(dtype) / 255
 			xp = torch.Tensor(x).type(dtype)/255
 
-		actions = torch.Tensor(actions,dtype=torch.long)
-		rewards = torch.Tensor(rewards,dtype=torch.float)
-		intrinsic_dones = torch.Tensor(intrinsic_dones,dtype=torch.uint8)
+		actions = torch.Tensor(actions)
+		rewards = torch.Tensor(rewards)
+		intrinsic_dones = torch.Tensor(intrinsic_dones)
 
 		if torch.cuda.device_count() > 0:
 			Qt.to(device0)
